Move KeycloakOidcUserGrantedAuthoritiesConverter out of servlet package
KeycloakOidcUserGrantedAuthoritiesConverter is used by both servlet/reactive client configurations
