+++ src/main/kotlin/com/github/daniel/shuy/oauth2/keycloak/client/servlet/KeycloakOidcUserService.kt
@@ -1,5 +1,6 @@
 package com.github.daniel.shuy.oauth2.keycloak.client.servlet
 
+import com.github.daniel.shuy.oauth2.keycloak.client.KeycloakOidcUserGrantedAuthoritiesConverter
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService
 import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser
